+++ tests/test_main_annotate.py
@@ -399,22 +399,28 @@ def test_annotate_implicit_style_filename(
     assert simple_file.read_text() == expected
 
 
-def test_annotate_unrecognised_style(fake_repository):
+def test_annotate_unrecognised_style(fake_repository, stringio):
     """Add a header to a file that has an unrecognised extension."""
     simple_file = fake_repository / "foo.foo"
     simple_file.write_text("pass")
 
-    with pytest.raises(SystemExit):
-        main(
-            [
-                "annotate",
-                "--license",
-                "GPL-3.0-or-later",
-                "--copyright",
-                "Jane Doe",
-                "foo.foo",
-            ]
-        )
+    result = main(
+        [
+            "annotate",
+            "--license",
+            "GPL-3.0-or-later",
+            "--copyright",
+            "Jane Doe",
+            "foo.foo",
+        ],
+        out=stringio,
+    )
+
+    assert result == 0
+    assert (
+        "foo.foo is not recognised; creating foo.foo.license"
+        in stringio.getvalue()
+    )
 
 
 def test_annotate_skip_unrecognised(fake_repository, stringio):
@@ -1265,32 +1271,6 @@ def test_annotate_recursive_on_file(fake_repository, stringio, mock_date_today):
         "Joe Somebody" in (fake_repository / "src/source_code.py").read_text()
     )
     assert result == 0
-
-
-def test_annotate_recursive_contains_unrecognised(
-    fake_repository, stringio, mock_date_today
-):
-    """Expect error and no edited files if at least one file has not been
-    recognised."""
-    (fake_repository / "baz").mkdir(parents=True)
-    (fake_repository / "baz/foo.py").write_text("foo")
-    (fake_repository / "baz/bar.unknown").write_text("bar")
-    (fake_repository / "baz/baz.sh").write_text("baz")
-
-    with pytest.raises(SystemExit):
-        main(
-            [
-                "annotate",
-                "--license",
-                "Apache-2.0",
-                "--copyright",
-                "Jane Doe",
-                "--recursive",
-                "baz/",
-            ]
-        )
-
-    assert "Jane Doe" not in (fake_repository / "baz/foo.py").read_text()
 
 
 # REUSE-IgnoreEnd